requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

flask~=1.1.2
pillow~=8.3.1